Reworked the Parameter class to contain a reference to its AbstractNode, optional parent Parameter and provide an after() callback hook for the end of the timestep.
This code also modifies the licence module and classes to use the refactored Parameter. Licences are now no longer a special set of objects, but just subclasses of Parameter. The LicenceCollection has been converted to a general ParameterCollection with two specific subclases providing minimum and maximum of the set of Parameters.

Tests updated for AnnualLicence. However existing XML tests with licences now fail and have been marked as such. No tests currently for the ParameterCollection or its subclasses.