Add a name to Recorder and use RecorderIterator for access in Model.
This approach is similar to NodeIterator but is more simple. RecorderIterator is a wrapper around a list of recorders but provides access through keys and also iteration. The used API for Model.recorders is not significantly changed, but now provides access via name.

Recorders have a default name that is the the lowercase of the class name. NodeRecorder and StorageRecorder provide a default naming include the node's name.