Add ParameterRecorder and IndexParameterRecorder.
This commit adds Recorder subclasses for Parameter and IndexParameter. Currently there are NumpyArray... versions of these base classes implemented and tested. Parameters now have access to the _recorders list to find out which Recorder objects are looking at them.

Spotted a couple of bugs on ControlCurveIndexParameter as well that are fixed.

TODO
- Currently there is no way to give a Parameter a name via keyword arguments to __init__.
- More subclasses to use the new Recorders.