Big refactor of Recorder and Parameter.
Both now share a common parent - Component - which is used to to determine the inheritance tree in order to call setup, reset, before, after and finish in a dependency tree. This removes the need for PairedSet. The biggest update to the API is the requirement to pass a `Model` instance to all Parameter objects, similar to the Recorder API. Many tests had to be updated to work with this.

In addition the Recorder.save method is now Recorder.after to be inline with the other objects.

This is still WIP and the dependency tree needs to be hit with some better tests.