Added Model.find_orphaned_components method. Closes #444. (#446)
* Added Model.find_orphaned_components method. Closes #444.

This commit adds a method helpful in searching for parameters in
the model that aren't used. AbstractNode now provides a components
property which yields all of the components it uses. This is done
using a class attribute, component_attrs, which lists all of the
properties on a class which may hold a component (e.g. max_flow).
This will allow nodes to define new properties in the future,
although most will be caught anyway as all nodes in the model
are queried, including the children of complex nodes.

* Fixed outdated comment in previous commit

* Removed initial_volume from Storage.component_attrs

* Raise warning if orphaned parameters in Model.check

* Fixed typo in previous commit