Add platypus optimisation support.
- Implements a new platypus model wrapper.
- Refactors the optimisation example to use either library.

The refactored approach wraps the platypus.Problem and creates a module
level cache of the Model for sub-processes to use. These module level values
are accessed through properties on the `BaseOptimisationWrapper`. This is parent
class shared by both the platypus and inspyred wrappers.

Also added tests for running an algorithm with a ProcessPoolEvaluator.