Refactored the schemas into a separate subpackage: pywr.schema
This subpackage will eventually contain schema for nodes, parameters and recorders. For now
it only contains a schema for the nodes, but it provides the basic implementation. The benefit of the
sub-package is that it means the schema do not need to be registered with each class. This lowers
the barrier of entry for new nodes and means the marshmallow code is not so tightly integrated.

Current loading procedure attempts to look up a schema based on the class name (JSON "type") of the node
being loaded. If such a schema does not exist an exception will be raised.

NB requires the development version of marshmallow.