A better workaround for #470
This change the reservoir volume initialisation in the case where
max_volume is parameter. Now both `initial_volume` and
`initial_volume_pc` must be specified as floats, and are used as
is. This removes the current whitelist of acceptable parameters.

It does however mean that if the user specifies initial conditions
that are not compatible with the actual max_volume (i.e.
initial_volume != initial_volume_pc/max_volume) then there is the
possibility that parameter values will be incorrect on the first
time-step. This changes effectively makes that the users
responsibility.