Fix loading PiecewiseLink with parameters from JSON.
The load method on PiecewiseLink was broken when given parameters. It
was also a little ambiguous regarding the use of 'max_flow' or 'max_flows',
and 'cost' or 'costs'. This is now made explicit to use the singular
case. That is consistent with the init methods of the class. Plural keys
are no longer supported.

Add a test for loading parameters on to PiecewiseLink sublinks.