Allow use of parameters as values in ControlCurveInterpolatedParameter
This copies the approach of ControlCurveParameter and allows the user
to provide either values as an iterable of floats or parameters as
an iterable of Parameter instances. Each parameter's value is used
during the calculation in the same way as the current list of values.

Tests extended to check this behaviour and support loading from JSON.