Added UniformDrawdownProfileParameter (#836)
* Added UniformDrawdownProfileParameter

This commit adds a new parameter that returns a recurring profile
that drawns down from 1.0 to 0.0 over the course of a year. The day
on which to start the drawdown is configurable with the same keywords
as to AnnualVirtualStorage.

Basic tests included. Also added a more complex AnnualVirtualStorage
test model that uses the new UniformDrawdownProfileParameter as
a control curve in a dynamic cost calculation.

* Fix typo in UniformDrawdownProfileParameter docstring.