Add `ControlCurvePiecewiseInterpolatedParameter` (#857)
* Add `ControlCurvePiecewiseInterpolatedParameter`

A new parameter the interpolates between multiple control curves in
a piecewise fashion. A pair of values is given for each "band"
between the storage bounds and the control curves. This pair is
used to set linear interpolation based on the current storage.

* Fix docstring typo in ControlCurvePiecewiseInterpolatedParameter
and check the dimensions of `values` inside the property setter
rather than `__init__` function.

* Add DeprecationWarning to PiecewiseLinearControlCurve.