Rework the Recorder constraint functionality to be more flexible.
Upper and lower bounds are now specified on Recorders instead of a
binary `is_constraint` flag. This allows the optimisaiton wrappers
to handle equality, single bounded and double bounded constraints.
It also allows the user to specify the constraint threshold. The
previous implementation relied on the default setting of the
respective optimisation wrappers (which weren't consistent).