Rework the Recorder constraint functionality to be more flexible. (#869)
* Rework the Recorder constraint functionality to be more flexible.

Upper and lower bounds are now specified on Recorders instead of a
binary `is_constraint` flag. This allows the optimisaiton wrappers
to handle equality, single bounded and double bounded constraints.
It also allows the user to specify the constraint threshold. The
previous implementation relied on the default setting of the
respective optimisation wrappers (which weren't consistent).

* Fix Recorder docstring re new bounds kwargs.

* Fix error in pygmo wrapper.

* Add constrained problem to pygmo tests.

* Add constraint violation and model feasibility methods.

A new method Recorder.is_constraint_violated() that computes
whether the value from the current simulation violates the
defined constraints. A Model.is_feasible() method returns
true if all constraints are not violated.

* Fix internal representation of undefined constraint upper bounds.

* Address review re docstrings and comments.

* Extend Model.is_feasible() docstring.