Add support for optimising the days of year in RbfProfileParameter.
Support optimising the days of year used for interpolation in
RbfProfileParameter. Support is provided using a new keyword
`variable_days_of_year_range` which defines the maximum deviation
from the given days of year that the optimisation use.

In addition new keywords, `min_value` and `max_value`, are added
which allowing capping the overall interpolated profile.