Refactored dynamic aggregated node factors.
Simplified the internal code by always updating the factors'
cefficient in the constraint matrix. This is inefficient when
the factors are constants, but means the API is consistent
with other `get_xxx` methods. The solver code is also
simplified with no distinction between fixed and dynamic
factors.

Also removed a now deprecated skipped test.