Implement a 2-stage loading scheme for nodes.
The changes here remove the existing `load` classmethod for
nodes. These are replaced with a new classmethod `pre_load`
which performs the object instantiation. After this itial
stage the parameters and recorders are fully loaded. Finally,
any parameter references are assigned to node attributes.

This resolves the circular loading between nodes and parameters.
However, it has required somworking of node init methods
which has impacted on the JSON schema for the nodes. Overall
this should make the JSON schema and init method arguments
more consistent.

Fixes #380.