Implement a 2-stage loading scheme for nodes. (#945)
* Implement a 2-stage loading scheme for nodes.

The changes here remove the existing `load` classmethod for
nodes. These are replaced with a new classmethod `pre_load`
which performs the object instantiation. After this itial
stage the parameters and recorders are fully loaded. Finally,
any parameter references are assigned to node attributes.

This resolves the circular loading between nodes and parameters.
However, it has required somworking of node init methods
which has impacted on the JSON schema for the nodes. Overall
this should make the JSON schema and init method arguments
more consistent.

Fixes #380.

* Re-instate Model._get_node_from_ref

To aid transition to the new node loading functional
it`_get_node_from_ref` is retained and defers to `model.nodes[]`
while issuing a DeprationWarning. Custom parameters that require
references to nodes will use this function during loading, but
should be updated in the future.

* Add Loadable to AggregatedStorage.

* Add test for RVS from JSON. Fix bug with failing to load.

* Remove unused __deferred_parameter_list_attributes__

* Add cost property to RiverGauge.