Add NormalisedGaussianKDEStorageRecorder
Works in a similar way to GaussianKDEStorageRecorder, but
derives from a new array recorder. NormalisedGaussianKDEStorageRecorder
keeps an array of storage that is normalised between -1 and 1
where the 0 value is at a control curve. The new KDE recorder
estimates a probability of being at or below zero in the normalised
series.