Selective bounds and objective updates for the glpk-edge solver (#912)
* Initial commit of more selective update of flow bounds.

Adds an optional setting to the glpk-edge based solver that,
when active, only sets constraint bounds for nodes with fixed
min_flow and max_flow values once. This happens during setup,
other nodes with parameters for either min_flow or max_flow
are updated as usual.

* Initial commit of a more selective update of costs.

Adds an optional setting to the glpk-edge based solver that,
when active, only gets costs from nodes that have a non-fixed
cost attribute. The fixed element of the costs is pre-calculated
in the setup if the model.

* Improve efficiency of fixed costs assignment.

* Remove typo in has_fixed_cost docstrings.

* Clarify glpk-edge comment re costs calculations.

* Add support for passing new glpk-edge parameters by environment variables.

* Improve fixed cost and flow constraint perf improvements.

- Improvements applied to regular route based GLPK solver.
- Solver settings reported to ModelResult object.
- Settings tested in CI and tests modified to work with
new settings.

* Address comment typos.