Update constant aggregated node factors only during rest. (#985)
* Update constant aggregated node factors only during rest.

This is a performance improvement that follows on from #983
by only updating aggregated node factors in the GLPK solvers
during reset instead of every timestep. It should have most
benefit for models which use many of aggregated nodes with factors
as the code both for calculating the normalised factors is
not very efficient.

A new solver option and corresponding environment variable is
added to turn this feature on. It is off by default.

Modified the CI test suite to run only with and without the
extra options. As opposed to running each option individually.

* Fix bug with glpk route solver updating all aggregated factors every timestep.

* Add set_fixed_factors_once to solver settings.