Assert entries when creating rows in the GLPK matrix.
In response to #1001 this commit implements a check to the
`glp_set_mat_row` call to ensure that there is at least one
entry in the row. Zero entry rows are likely caused by invalid
or unsupported model / network configuration (e.g. those
reported in #1001). For the time being this change will ensure
such models fail to run with an `AssertionError` instead of
failing to applying expected constraints silently.

A proper fix to enable support for a `PiecewiseLink` or other
complex nodes with `VirtualStorage` is still required.