Improve GLPK error handling. (#1021)
* Improve GLPK error handling.

This picks up some work from #759 and the work done in #762 to
improve the error handling in GLPK. Best to read the new page
in the documentation that's been added about this. The TLDR
is that this adds NaN checks and GLPK error handling to Pywr
by default at the cost of some performance. The older
behaviour can be created with `--glpk-unsafe` compile time flag.

There are two new tests that cause seg faults with the unsafe
interface (i.e. current behaviour). Raising exceptions is better
from a handling point of view for users wrapping models in
multi-processing, etc.

Queries:
 - Should the unsafe interface be a runtime option instead of
a compile time option?
- Is changing the default error handling behaviour a good idea?

* Fix format of setup.py

* Actually add the new documentation.

* Skip NaN tests for lpsolve.

* Fix test formatting.

* Fix except instead of except?

* Remove print statements.

* Tweaks to docs in response to review comments.

* Handle exceptions from GLPK low-level methods.

This should ensure the setup methods do not leak any memory
and that a useful error message is logged.

* Make unsafe GLPK API a runtime option.

* Remove unsafe flags from setup.py

* Fix bug in unsafe API and skip some tests with unsafe API.

* Correct keyword argument name for using unsafe GLPK API.