Correct the use of keyword arguments in Model.load (#1043)
* Correct the use of keyword arguments in Model.load

This fixes a bug with how keyword arguments are passed between
the various load methods. The bug meant that any solver passed
to .load() was ignored. This fixes that and adds tests. The
keyword argument overrides both the solver settings in the JSON
data and PYWR_SOLVER environment variable.

Fixes #1042

* Address test failures and fix test capture.