fix: Overload get_xxx_flow for PiecewiseLink.
This fix allows PiecewiseLink to return the sum of its sublinks
min and max flows. While the solver won't use these methods
because the outer node is not added to the LP it does allow
recorders to get the total min and max flows for the entire
node.

Test implemented for deficit recorder.

Fixes #1087.